requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.16
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.26.0
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
slowapi==0.1.9
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
